Consolidate integration types
This commit replaces the hierarchy of different IntegrationType subclasses
for every integration type with a smaller hierarchy containing one class
for position integration, another for momentum integration, etc. Different
dimensionalities of integration (dipole vs. quadrupole etc.) are now represented
by different instances of the same class.

To enable this we need a comparator for IntegrationType instances that is more
intelligent than comparing based on memory address, because now the constructors
are public and so there can be multiple instances with the same properties that
should compare equal. The signature of HardFactorTypeMap is updated accordingly.