Introduce new polar integration types
Previously the RadialIntegrationType assumed that the hard factors it
was being used for had no angular dependence, and thus when given a
radial coordinate r it would always set rx=r,ry=0. This commit
introduces an integration type that does allow for angular dependence,
and properly sets rx=r cos(theta),ry=r sin(theta). It also introduces an
analogous integration type for momentum, and extends that type to one
that provides an upper cutoff on q_perp. The cutoff is calculated to
satisfy x_a <= 1 in the exact kinematics.