Get IntegrationType to calculate the jacobian
Fully specifying an integral requires specifying the bounds, the
transformations from the variables of integration to the kinematic
variables, and the jacobian. With this commit, all of this is handled
in the IntegrationType, rather than the jacobian being calculated
in integrator.cpp, so the jacobian can change with the integration
type as it should.

Conflicts:
	src/integrationtype.h