Move the conversion of (z,y) to (z,xi) into IntegrationType instances
This commit moves the code that converts from the integration variables
(z,y) to the kinematic variables (z,xi) from
IntegrationContext::update_kinematics() to IntegrationType::update().
This way the IntegrationContext object only "knows" about the kinematic
variables. Everything to do with converting to and from Monte Carlo
integration variables with constant limits is now encapsulated in the
methods of IntegrationType.