Create separate functions to calculate kinematic limits
This commit breaks out the computation of the kinematic limits on z and
y into separate functions in integrationtype.cpp. These limits can be
easily adjusted by altering the code in just these few functions. It
also reduces the number of places in which the exact_kinematics flag
is referenced.