Fix bug with double-including jacobian
QLimitedMomentumIntegrationType's momentum coordinates alternate
between radii and angular variables, so extra_dimensions is twice
the number of radial coordinates being integrated over. Only the
radial coordinates are supposed to have a nontrivial jacobian, but
in previous versions, the jacobian contained a factor of qmax for
each coordinate, radial or angular. This commit fixes that, so that
only the radial coordinates contribute a factor of qmax to the
jacobian.