Allow specifying std or hipt diagonal terms
This commit adds the additional hard factor group shortcuts "nlo.std"
and "nlo.hipt" to manually specify whether to use the original hard
factors for the NLO diagonal channels, or the high-pT expanded ones
from the matching paper (1405.6311). The shortcut "nlo" is now a
meta-shortcut that acts like either of those depending on whether
exact kinematics are in force.

The README file has been updated to reflect this change.