Add MuParser and gslmuparser as dependencies
MuParser is a library to parse mathematical expressions. gslmuparser
is a wrapper around GSL functions that allows them to be used in MuParser.
These are being added in preparation for creating the hard factors by
reading and parsing a text file at runtime, instead of needing to
write their code by hand.