Fix bug where hard factors would end prematurely
Prior to this commit, when parsing a hard factor definition file, if
multiple functions (e.g. Fs and Fd) were defined for a single hard factor
then only the first would be associated with that hard factor; the next
would be associated with the following hard factor because the program
thought the hard factor definition was complete once it encountered
a single function definition. This commit fixes that. Now the parser
will wait until it actually finds a repeat property before finalizing
the hard factor definition.