Remove config files from run/ and put hf spec file in source tree
This commit removes the files that were added to the run/ directory,
because that directory should be left clean for local run configurations.
The hard factor specification file hardfactors.txt has been tested, verified
to agree with the direct (non-parsed) implementation, and thus it has
been moved into the source tree as src/hardfactors/hardfactors.cfg.
An installation rule was added to src/hardfactors/CMakeLists.txt
to install this file.