Create an integration type to integrate over r/xi
The new integration type is a functional duplicate of
AngleIndependentPositionIntegrationType except that its position
variable of integration is scaled by a factor of 1/xi. In other words,
if we define rperp' = rperp/xi, the new integration type integrates over
rperp' instead of rperp. The point of this is to reduce the amount by
which kperp*rperp/xi varies, since this factor is the argument of a
Bessel function and reducing its variation will lead to a smoother
integrand. Hopefully, anyway.