New HardFactorRegistry impl that allows arbitrary implementation codes
This commit rewrites the HardFactorRegistry implementation so that any
arbitrary string can be used as an implementation code, not just "m",
"r", and "p". It also creates a common data structure, category_map,
that allows accessing elements by name and implementation code or just
by name. This sets up for future commits to use implementation codes
for HardFactorGroups.

This commit also cleans up the implementation of hard factor storage.