Add memory management to registry and category map
This commit adds copy constructors and assignment operators and
the like to HardFactorRegistry and category_map. It also changes
the HardFactorParser to hold a reference to a registry, not a copy,
so that any hard factors added through the parser go back into the
original registry (thus fixing a bug).