Defer parsing of hard factor groups and use last hard factor by default
This commit changes the behavior of HardFactorParser so that whenever
a group specification is read from the file, instead of being immediately
parsed as before, it is now stored until the (new) flush_groups() method
is called. This method goes over all stored group specifications and
parses them using whatever hard factors are entered into the registry
at that time. If a HardFactorParser is deleted or goes out of scope,
flush_groups() will be automatically called from the destructor.

Making this change enables hard factor definition files to define a group
group:hfa,hfb
and specify implementations of hfa and hfb later in the file, or even in
a different file to be loaded afterwards. This also means a file loaded
later in the process can offer alternative definitions of hard factors
that will override the earlier definitions in group specifications.

Oh yeah, this commit also changes category_map so that the get(name)
method chooses the _last_ item to be added with the given name, not
the _first_ as was previously the case.