Update hard factor defs to put momentum hfs last
After the last commit, the last hard factor definition with a given name
is now treated as the default. This commit reorders the definitions in
the included files so that the momentum-space hard factors are always
later than the position-space ones, so that they will continue to be
the defaults.