Hide seed information if running with one seed
This commit causes all the extra output added in the last couple commits,
which relates to multiple pseudorandom seeds (e.g. the seed column in
the output, the mean and standard deviation across seeds) to be hidden
if the program is running with only one seed. This way the data processing
scripts don't have to be changed, as long as they are only given outputs
corresponding to single-seed runs.