Change parsing so blank lines delimit hard factors
This commit changes the parsing scheme for hard factor definitions
such that a definition ends at a blank line (or at the end of the
file), rather than ending when a repeated property is specified
as was the case before. It also improves the parsing infrastructure
in some minor ways.