Switch from separate HardFactorList* and string to HardFactorGroup*
This commit changes oneloopcalc.cpp and other files to use a vector of
HardFactorGroup pointers, instead of separate vectors of HardFactorList
pointers and strings for their names. This simplifies the code (it makes
more sense to just use the object that already has the hard factors and
the group label together) and reduces use of pointers; in some places
they've been swapped out for references to help with memory management.