Explicitly initialize all fields in FileDataGluonDistribution
Apparently the logic that decided which fields in FDGD needed to be deallocated
was really bad, and it would wind up freeing memory that hadn't actually been
allocated by the class in the first place. This commit explicitly initializes
all the fields to NULL so that the various ...free() calls don't corrupt the
memory if they are used on something that wasn't initialized.

A real fix would also ensure that everything which does get allocated is freed,
but I'm not being that careful right now. So there may still be memory leaks,
but they should be minor.

Also let the record reflect that this is the bug which caused the term
"Schroedingerror" to come into being.