Separate config parsing from Context management and split oneloopcalc.cpp
This commit creates a separate class (and source files) to just read a
configuration file and store the key/value combinations within. The config
file reader doesn't do any parsing or interpreting of the values.
That is now the sole purvey of ContextCollection.

It also splits oneloopcalc.cpp into separate files for the ProgramConfiguration,
ResultsCalculator, and the main program itself, along with a separate header for
the trace variable stuff. There are also some changes to enforce
const-correctness a little more consistently. These are all auxiliary changes
that make it easier to manage the impact of having the configuration parsing
separate from the Context creation.