Limit q2maxMV to 1e4 by default to avoid numerical instabilities
For extreme conditions when the phase space is very large (looking
at you FCC), q2max becomes so large that the numerical Fourier transform
of the MV distribution fails to converge. In that case, as of this
commit, the code limits q2maxMV to 1e4 by default, which is small enough
not to encounter numerical problems in typical runs, but large enough
to be in the 1/q^4 scaling regime. This is just a default; the actual
value should be tuned from run to run.