Fix bug with passing Qs2 to gluon distribution
The Fourier transform implementation in AbstractTransformDistribution
would pass Qs2 as the second parameter to F() or S(), although that
parameter is supposed to be rapidity. This causes the computed
transformed gluon distribution to be very incorrect any time Qs2
is not numerically close to Y. This commit fixes that bug.