Preliminary conversion to IntegrationRegion instead of IntegrationType
This commit replaces the preexisting IntegrationTypes with
IntegrationRegions, which can be composed together. All aspects of the
integrals (limits, integration variables, jacobians) are now encapsulated
by the IntegrationRegion class. There is no more switching variables
based on whether the hard factor is leading or next-to-leading order,
or based on approximate or exact kinematics.

This change also requires major changes to the parser to allow
for a new syntax of specifying integration regions in the file,
and some changes to the hard factor implementation and the integrator,
though these are minor because the essential functionality from
IntegrationType (min, max, jacobian, update) is still there.

The code should compile as of this commit, but still contains logic errors.